feat: MovieList 컴포넌트 구현
