+++ src/components/MovieChart/index.ts
@@ -1,7 +1,7 @@
 import { getPopularMovies, GetPopularMoviesRes, waitFor } from './../../apis/index';
 import { assemble, Event, useEffect, useState } from '../../core';
 import { getElement } from './../../utils/common/domHelper';
-import { MovieComponent } from './action';
+import { MovieListComponent } from './action';
 
 const MovieChart = assemble(() => {
   const [movieList, setMovieList] = useState<GetPopularMoviesRes | null>(null);
@@ -20,14 +20,9 @@ const MovieChart = assemble(() => {
       <section class="item-view">
         <h2>지금 인기 있는 영화</h2>
         <ul class="item-list">
-        ${movieList?.results
-          .map(
-            (info) => `
-          <fragment id="Movie-${info.id}">
-            ${MovieComponent({ info })}
-          </fragment>`
-          )
-          .join('')}
+          <fragment id='MovieList'>
+            ${movieList ? MovieListComponent({ movieList: movieList.results }) : ''}
+          </fragment>
         </ul>
         <button class="btn primary full-width">더 보기</button>
       </section>
