+++ src/components/MovieList/action.ts
@@ -1,0 +1,16 @@
+import { debounce } from './../../utils/common/debounce';
+import { $, replaceComponent } from './../../utils/common/domHelper';
+import { Movie, MovieProps } from '../Movie';
+
+const MovieListActions = () => {
+  const MovieComponent = (props: MovieProps) => {
+    return (
+      debounce(() => {
+        replaceComponent($(`#Movie-${props.info.id}`), Movie(props));
+      })() && ''
+    );
+  };
+  return { MovieComponent };
+};
+
+export const { MovieComponent } = MovieListActions();
